BCryptPasswordHandling class created, methods done but need to test it, Not sure if hashing is working.

+++ src/main/java/com/codecool/petproject/newspostal/module/User.java
@@ -10,7 +10,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public double id;
+    public long id;
 
     public String email;
